+++ facade/src/test/scala/react/semanticui/collections/menu/MenuMenuTests.scala
@@ -1,0 +1,23 @@
+package react.semanticui.collections.menu
+
+import utest._
+import japgolly.scalajs.react.test._
+
+object MenuMenuTests extends TestSuite {
+  val tests = Tests {
+    'render - {
+      val menuItem = MenuMenu(MenuMenu.props())
+      ReactTestUtils.withNewBodyElement { mountNode =>
+        menuItem.renderIntoDOM(mountNode)
+        assert(mountNode.outerHTML == """<div><div class="menu"></div></div>""")
+      }
+    }
+    'renderAs - {
+      val menuItem = MenuMenu(MenuMenu.props(as = "a"))
+      ReactTestUtils.withNewBodyElement { mountNode =>
+        menuItem.renderIntoDOM(mountNode)
+        assert(mountNode.outerHTML == """<div><a class="menu"></a></div>""")
+      }
+    }
+  }
+}